requirements.txt
Django==4.2.11
gunicorn==21.2.0
python-dotenv
dj-database-url
psycopg2-binary
django-bootstrap5
whitenoise
django-filter
